Support choosing from a collection of scaffolds (#3)
* Support choosing from a collection of scaffolds

Add the option to override variables as this helps with testing of
collections.

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Prefer overrides to passing in io.ReadCloser

Given the support for overrides, we can use overrides in testing as
opposed to passing in io.ReadCloser.

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Skip input read when override exists

We do not need to read from the user if an override exists

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Support collections in the `scafall` binary

Split tests into integration (testing the pkg) and system (testing the
scafall binary).

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Introduce internal package to improve unit testing

Use an internal package so we can expose some internal functions for
unit testing while not making them part of the package's public
interface

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Test only whether a file is executable by user

We can not rely on the UMASK when writing files to
the filesystem.  Check only the u+x permissions.

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Improve project documentation

Provide examples of API usage and better top-level documentation for
the package.

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>