Remove concept of defaults (#8)
* Remove collections and defaults concepts

Remove the idea of template collections.  Also remove the idea of
defaults.  We can revisit this design decision in future.

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Fix output-folder check in rootCmd

Check to see if an error occurred

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Add command line overrides

Use --override or -o to provide key=value overrides

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Improve documentation on overrides

Drop all discussion of DefaultValues from documentation and note that
prompts.toml and root-level README are not part of the generated project.

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Support scaffolding from a sub directory

Template projects may provide scaffolding in a sub-directory.  Allow
the user to specify the subPath they wish to use inside the template
project.

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>

* Re-introduce the idea of a collection

A collection is now simply a set of subPath's inside a project.

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>