Clean up module interface (#10)
* Get list of arguments for template

Each template defines a list of arguments or, the template collection
defines a list of templates.  Allow scafall to detect and inform the
user of the available options.

* Rename Overrides to Arguments

The term Arguments is clearer and easier to explain to end users.
Prefer this term over the custom term.

* Clean up interface

Hide implementation details in the interface such as the tmpDir and
whether the repo has previously been cloned.

Signed-off-by: Aidan Delaney <adelaney21@bloomberg.net>